+++ apps/meteor/ee/client/views/admin/engagementDashboard/EngagementDashboardRoute.tsx
@@ -1,4 +1,4 @@
-import { usePermission, useRouter, useSetModal, useCurrentModal, useTranslation } from '@rocket.chat/ui-contexts';
+import { usePermission, useRouter, useSetModal, useCurrentModal, useTranslation, useRouteParameter } from '@rocket.chat/ui-contexts';
 import type { ReactElement } from 'react';
 import React, { useEffect } from 'react';
 
@@ -21,12 +21,13 @@ const EngagementDashboardRoute = (): ReactElement | null => {
 	const isModalOpen = useCurrentModal() !== null;
 
 	const router = useRouter();
-	const { tab } = router.getRouteParameters();
+	const tab = useRouteParameter('tab');
 	const eventStats = useEndpointAction('POST', '/v1/statistics.telemetry');
 
 	const hasEngagementDashboard = useHasLicenseModule('engagement-dashboard') as boolean;
 
-	const { shouldShowUpsell, cloudWorkspaceHadTrial, handleGoFullyFeatured, handleTalkToSales } = useUpsellActions(hasEngagementDashboard);
+	const { shouldShowUpsell, cloudWorkspaceHadTrial, handleManageSubscription, handleTalkToSales } =
+		useUpsellActions(hasEngagementDashboard);
 
 	useEffect(() => {
 		if (shouldShowUpsell) {
@@ -37,7 +38,7 @@ const EngagementDashboardRoute = (): ReactElement | null => {
 					subtitle={t('Analyze_practical_usage')}
 					description={t('Enrich_your_workspace')}
 					onClose={() => setModal(null)}
-					onConfirm={handleGoFullyFeatured}
+					onConfirm={handleManageSubscription}
 					confirmText={cloudWorkspaceHadTrial ? t('Learn_more') : t('Start_a_free_trial')}
 					onCancel={handleTalkToSales}
 					cancelText={t('Talk_to_an_expert')}
@@ -45,16 +46,18 @@ const EngagementDashboardRoute = (): ReactElement | null => {
 			);
 		}
 
-		if (!isValidTab(tab)) {
-			router.navigate(
-				{
-					pattern: '/admin/engagement/:tab?',
-					params: { tab: 'users' },
-				},
-				{ replace: true },
-			);
-		}
-	}, [shouldShowUpsell, router, tab, setModal, t, handleGoFullyFeatured, cloudWorkspaceHadTrial, handleTalkToSales]);
+		router.subscribeToRouteChange(() => {
+			if (!isValidTab(tab)) {
+				router.navigate(
+					{
+						pattern: '/admin/engagement/:tab?',
+						params: { tab: 'users' },
+					},
+					{ replace: true },
+				);
+			}
+		});
+	}, [shouldShowUpsell, router, tab, setModal, t, handleManageSubscription, cloudWorkspaceHadTrial, handleTalkToSales]);
 
 	if (isModalOpen) {
 		return <PageSkeleton />;